feat: rename bundlr to irys
