+++ packages/irys/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@hey/bundlr",
+  "name": "@hey/irys",
   "version": "0.0.0",
   "private": true,
   "license": "AGPL-3.0",
