+++ packages/workers/metadata/.dev.vars.example
@@ -1,2 +1,2 @@
-BUNDLR_PRIVATE_KEY=""
+IRYS_PRIVATE_KEY=""
 HUGGINGFACE_API_KEY=""
