+++ packages/workers/metadata/package.json
@@ -15,8 +15,8 @@
     "worker:deploy": "wrangler deploy --var RELEASE:\"$(git rev-parse HEAD)\""
   },
   "dependencies": {
-    "@hey/bundlr": "workspace:*",
     "@hey/data": "workspace:*",
+    "@hey/irys": "workspace:*",
     "@hey/lens": "workspace:*",
     "@hey/lib": "workspace:*",
     "itty-router": "^4.0.23"
