+++ script/bootstrap
@@ -16,7 +16,7 @@ echo "Updating packages/workers/metadata env vars 🌎"
 cd packages/workers/metadata
 # Generate private key and insert into .dev.vars
 random_hex=$(head -c 32 /dev/urandom | xxd -p -c 32)
-echo "BUNDLR_PRIVATE_KEY=$random_hex" > .dev.vars
+echo "IRYS_PRIVATE_KEY=$random_hex" > .dev.vars
 cd ../../../
 # -------------------------------------
 echo "Updating packages/workers/snapshot-relay env vars 🌎"
